Island: Move AuthenticationService to authentication directory
