+++ monkey/monkey_island/cc/services/__init__.py
@@ -1,0 +1,1 @@
+from .authentication.authentication_service import AuthenticationService
