+++ monkey/monkey_island/cc/services/authentication/authentication_service.py
