+++ monkey/monkey_island/cc/services/initialize.py
@@ -1,6 +1,7 @@
-from monkey_island.cc.services.authentication import AuthenticationService
 from monkey_island.cc.services.post_breach_files import PostBreachFilesService
 from monkey_island.cc.services.run_local_monkey import LocalMonkeyRunService
+
+from . import AuthenticationService
 
 
 def initialize_services(data_dir):
